SDL-87 Tooltip should have visibility:hidden when it is not open
By default, the Tooltip is hidden visually with CSS `opacity: 0`. This
means the element is still visible to assistive technology and can
potentially be interacted with using the mouse. It should have CSS
`visibility: hidden`, `display: none` or `hidden` attribute when not
shown.

Since we're using a CSS transition on `opacity`, `visibility: hidden`
is probably the simplest CSS-only solution.